requirements.txt
# Core dependencies
langchain>=0.1.0
langchain-community>=0.0.10
langchain-google-genai>=0.0.5
google-generativeai>=0.3.0

# Vector store and embeddings
faiss-cpu>=1.7.4
sentence-transformers>=2.2.2
langchain_huggingface>=0.0.5

# Web interface
streamlit>=1.31.0

# Environment variables
python-dotenv>=1.0.0

# Data processing
pandas>=2.1.0

# Testing
pytest>=7.4.0
flake8>=6.1.0